Move FieldSpec to separate module
Also, for *Value imports, use cimport since they're all defined in proper
cython modules. This improves performance *slightly* by avoiding some of the
Python overhead.
